feat: 역 이름 중복 및 노선 이름, 색상 중복 처리 구현

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -2,7 +2,9 @@ package wooteco.subway.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.common.exception.badrequest.DuplicateNameException;
+import wooteco.common.exception.badrequest.DuplicateStationNameException;
+import wooteco.common.exception.badrequest.NotRemovalStationException;
+import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.web.dto.request.StationRequest;
@@ -14,14 +16,19 @@ import java.util.stream.Collectors;
 @Service
 @Transactional(readOnly = true)
 public class StationService {
-    private StationDao stationDao;
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
+        if (stationDao.findByName(stationRequest.getName()).isPresent()) {
+            throw new DuplicateStationNameException();
+        }
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
@@ -40,6 +47,9 @@ public class StationService {
 
     @Transactional
     public void deleteStationById(Long id) {
+        if (sectionDao.existsStation(id)) {
+            throw new NotRemovalStationException();
+        }
         stationDao.deleteById(id);
     }
 
@@ -47,7 +57,7 @@ public class StationService {
     public StationResponse updateStation(Long id, StationRequest updateStationRequest) {
         String updateName = updateStationRequest.getName();
         if (stationDao.findByName(updateName).isPresent()) {
-            throw new DuplicateNameException();
+            throw new DuplicateStationNameException();
         }
         stationDao.update(id, updateName);
         return new StationResponse(id, updateName);
